discussion/requirements.txt
alembic==1.7.5
apispec==5.1.1
attrs==21.4.0
autopep8==1.6.0
click==8.0.3
coverage==6.2
Deprecated==1.2.13
flasgger==0.9.5
Flask==2.0.2
flask-apispec==0.11.0
Flask-JWT==0.3.2
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
flask-smorest==0.35.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==4.3.3
Mako==1.1.6
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-sqlalchemy==0.27.0
mistune==2.0.1
packaging==21.3
pycodestyle==2.8.0
PyJWT==1.4.2
pyparsing==3.0.6
pyrsistent==0.18.0
PyYAML==6.0
redis==4.1.0
six==1.16.0
SQLAlchemy==1.4.29
toml==0.10.2
webargs==8.0.1
Werkzeug==2.0.2
wrapt==1.13.3
